Moved pkg/osvscanner/customgitignore/ to internal/customgitignore/
