+++ internal/customgitignore/walk_up_to_root.go
